AdProduct /KP

+++ src/KABKABEhandel/Models/DAL/EHandelDB.cs
@@ -92,7 +92,7 @@ namespace KABKABEhandel.Models.DAL
                 product.Name = row["Name"].ToString();
                 product.ImageURL = row["ImageUrl"].ToString();
                 product.Vat = Convert.ToDouble(row["Vat"].ToString());
-                product.Price = Convert.ToDouble(row["Price"].ToString());
+                product.Price = Convert.ToDecimal(row["Price"].ToString());
                 product.IsActive = Convert.ToBoolean(row["IsActive"].ToString());
                 product.Description = row["Description"].ToString();
                 product.Discount = Convert.ToDouble(row["Discount"].ToString());
